refactor: delete stale files.

+++ gen.sh
@@ -1,6 +1,10 @@
-cd outputs
-rm -rf *
-cd ..
+if [ -d "/outputs" ]; then
+	cd outputs
+	rm -rf *
+	cd ..
+else
+	mkdir outputs
+fi
 
 rm main.r1cs
 rm -rf main_js
@@ -12,12 +16,12 @@ cd ..
 
 # Phase 1.
 snarkjs powersoftau new bn128 15 outputs/main.ptau -v
-snarkjs powersoftau contribute outputs/main.ptau outputs/main2.ptau --name="ATTP Default Contribution 2" -v
+snarkjs powersoftau contribute outputs/main.ptau outputs/main2.ptau --name="Abyss Testnet Contribution 1" -v
 
 # Phase 2.
 snarkjs powersoftau prepare phase2 outputs/main2.ptau outputs/main_final.ptau -v
 snarkjs groth16 setup main.r1cs outputs/main_final.ptau outputs/main.zkey
-snarkjs zkey contribute outputs/main.zkey outputs/main2.zkey --name="ATTP Default Contribution 3" -v
+snarkjs zkey contribute outputs/main.zkey outputs/main2.zkey --name="Abyss Testnet Contribution 2" -v
 snarkjs zkey export verificationkey outputs/main2.zkey outputs/verification_key.json
 snarkjs groth16 prove outputs/main2.zkey main_js/witness.wtns outputs/proof.json outputs/public.json
 snarkjs groth16 verify outputs/verification_key.json outputs/public.json outputs/proof.json
